requirements.txt
streamlit==1.41.0          # Optional: Streamlit prototype UI
google-generativeai==0.8.3 # Official Google Generative AI / Gemini Python SDK
python-dotenv==1.0.1       # Convenient local .env loading for GEMINI_API_KEY
pandas==2.2.3              # Tabular display for emotional arcs and risk predictions
fastapi==0.115.6           # Backend API framework
uvicorn[standard]==0.32.1  # ASGI server for FastAPI
pydantic==2.9.2            # Request/response validation models
